requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pdfplumber
scikit-learn
beautifulsoup4
requests
jinja2
python-dotenv